requirements.txt
transformers==4.56.2
peft==0.18.1
accelerate==1.12.0
bitsandbytes==0.49.2
datasets
trl==0.28.0
google-cloud-aiplatform
python-dotenv

torch
scipy
sentencepiece
protobuf
